requirements.txt
aiogram>=3.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0